backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
anthropic>=0.40.0
python-docx==1.1.2
beautifulsoup4==4.12.3
pytest>=9.0.0
httpx>=0.28.0
